requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
werkzeug==3.0.3
wtforms==3.1.2
